Moved MessageHelper impl from :feature:stopdetail to :core:ui-impl
