+++ app/build.gradle
@@ -50,4 +50,5 @@ dependencies {
 
     implementation project(":feature:stoplist")
     implementation project(":feature:stopdetail")
+    implementation project(":core:ui-impl")
 }
