+++ core/ui-impl/build.gradle
@@ -1,6 +1,8 @@
 plugins {
     alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
+    alias(libs.plugins.hilt)
+    alias(libs.plugins.kapt)
 }
 
 android {
@@ -30,6 +32,8 @@ android {
 }
 
 dependencies {
+    implementation libs.hilt.android
+    kapt libs.hilt.compiler
 
     implementation libs.androidx.core.ktx
     implementation libs.androidx.appcompat
@@ -37,4 +41,6 @@ dependencies {
     testImplementation libs.junit
     androidTestImplementation libs.androidx.junit
     androidTestImplementation libs.androidx.espresso.core
+
+    implementation project(":core:ui")
 }
