+++ core/ui-impl/src/main/java/com/chaeny/busoda/uiimpl/di/MessageHelperModule.kt
@@ -1,6 +1,7 @@
-package com.chaeny.busoda.stopdetail
+package com.chaeny.busoda.uiimpl.di
 
 import com.chaeny.busoda.ui.MessageHelper
+import com.chaeny.busoda.uiimpl.ToastMessageHelper
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
