+++ feature/stopdetail/build.gradle
@@ -53,5 +53,4 @@ dependencies {
 
     implementation project(":core:data")
     implementation project(":core:model")
-    implementation project(":core:ui")
 }
